Moves some utility methods to features/base/devices.
