+++ lang/main.json
@@ -477,7 +477,9 @@
         "somebody": "Somebody",
         "suboptimalExperienceDescription": "Eer... we are afraid your experience with __appName__ isn't going to be that great here. We are looking for ways to improve this but, until then, please try using one of the <a href='static/recommendedBrowsers.html' target='_blank'>fully supported browsers</a>.",
         "suboptimalExperienceTitle": "Browser Warning",
-        "newDeviceTitle": "New __deviceType__ detected",
+        "newDeviceCameraTitle": "New camera detected",
+        "newDeviceMicTitle": "New microphone detected",
+        "newDeviceCameraTitle": "New audio output detected",
         "newDeviceAction": "Use"
     },
     "passwordSetRemotely": "set by another member",
